% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_srcpkgs.R
\name{find_srcpkgs_paths}
\alias{find_srcpkgs_paths}
\title{find all source packages inside a directory}
\usage{
find_srcpkgs_paths(dir, src = TRUE, prune = TRUE, ...)
}
\arguments{
\item{dir}{the directory in which to find the pkgs}

\item{src}{only report source packages, not installed ones}

\item{prune}{whether to discard pkgs inside other pkgs}

\item{...}{for \code{\link{list.files}}}
}
\value{
a vector of (relative) package paths
}
\description{
by default not in hidden (.) dirs, cf list.files() all.files= param
}
\keyword{internal}
