% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_root.R
\name{get_project_root}
\alias{get_project_root}
\title{get the current project root}
\usage{
get_project_root()
}
\value{
the project root dir path as a string, or NULL if none
}
\description{
get the current project root
}
