% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_r_loaders.R
\name{hack_r_loaders}
\alias{hack_r_loaders}
\title{instruments the base library() and loadNamespace() to make them aware of source packages}
\usage{
hack_r_loaders()
}
\description{
hacks library() and loadNamespace() using trace()
library(package) will basically call pkg_load(package) if package is managed by srcpkgs
}
\seealso{
unhack_r_loaders
}
