% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_root.R
\name{init_project_root}
\alias{init_project_root}
\title{initializes the project root if needed and checks it}
\usage{
init_project_root(root = find_project_root())
}
\arguments{
\item{root}{the project root dir path to set if no root is already set}
}
\value{
the project root invisibly
}
\description{
initializes the project root if needed and checks it
}
