% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{pkg_create}
\alias{pkg_create}
\title{creates and populates a R package-like folder programmatically, useful for writing tests}
\usage{
pkg_create(
  dir,
  name,
  functions = list(dummy = function() "DUMMY"),
  imports = NULL,
  depends = NULL,
  suggests = NULL
)
}
\arguments{
\item{dir}{the directory in which to create the package, as a string}

\item{name}{the package name, as a string}

\item{functions}{a named list of functions to add to the package}

\item{imports}{the "imports" dependencies}

\item{depends}{the "depends" dependencies}

\item{suggests}{the "suggests" dependencies}
}
\value{
the srcpkg instance, invisibly
}
\description{
basically a wrapper around \code{utils::package.skeleton()}
}
\keyword{internal}
