% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pkg_deps}
\alias{pkg_deps}
\title{computes the dependencies of a (source) package}
\usage{
pkg_deps(
  pkgid,
  src_pkgs = find_srcpkgs(),
  source = TRUE,
  installed = TRUE,
  imports = TRUE,
  depends = TRUE,
  suggests = TRUE,
  reverse = FALSE
)
}
\description{
computes the dependencies of a (source) package
}
