% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_roxygenise.R
\name{pkg_roxygenise}
\alias{pkg_roxygenise}
\title{roxygenize a source package if needed}
\usage{
pkg_roxygenise(pkg_path, force = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{force}{if force(d), do not use the md5-based system to detect package changes}

\item{quiet}{whether to be quiet/silent}

\item{...}{passed to  \code{\link[devtools:document]{devtools::document()}}}
}
\value{
if the roxygenation has been performed
}
\description{
\itemize{
\item if the package has not changed (based on the md5sum file), does nothing
\item otherwise roxygenise the package using roxygen2::roxygenise
\item and update and save the new md5sum file
}
}
\details{
\itemize{
\item N.B: has the side-effect of loading the package
}
}
\keyword{internal}
