% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_test.R
\name{pkgs_test}
\alias{pkgs_test}
\title{tests a list of packages}
\usage{
pkgs_test(
  pkgids = names(src_pkgs),
  src_pkgs = get_srcpkgs(),
  quiet = TRUE,
  fail_on_error = FALSE,
  ...
)
}
\description{
tests a list of packages
}
