% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_root.R
\name{set_project_root}
\alias{set_project_root}
\title{set the current project root}
\usage{
set_project_root(path, force = FALSE)
}
\arguments{
\item{path}{the project root dir path}

\item{force}{whether to bypass sanity checks and set the value}
}
\value{
the previous root as a string, or NULL
}
\description{
set the current project root
}
