% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_srcpkgs.R
\name{set_srcpkgs}
\alias{set_srcpkgs}
\title{set the current list of source packages}
\usage{
set_srcpkgs(src_pkgs)
}
\arguments{
\item{src_pkgs}{a collection of source packages as a \code{srckgs} object.}
}
\value{
the previous registered value (may be NULL) invisibly
}
\description{
set the current list of source packages
}
