% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_srcpkgs.R
\name{set_srcpkgs_paths}
\alias{set_srcpkgs_paths}
\title{set the current paths of source packages}
\usage{
set_srcpkgs_paths(paths)
}
\arguments{
\item{paths}{the paths of source packages to register}
}
\value{
the previous registered value (may be NULL) invisibly
}
\description{
set the current paths of source packages
}
