% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{setup_and_get_dummy_srcpkg}
\alias{setup_and_get_dummy_srcpkg}
\title{installs the dummy srcpkg in a temp location}
\usage{
setup_and_get_dummy_srcpkg(dest = tempfile())
}
\arguments{
\item{dest}{where to install the dummy srcpkg}
}
\value{
the package as a \code{srcpkg} object
}
\description{
Intended for testing and to write examples
}
\examples{
pkg <- setup_and_get_dummy_srcpkg()
print(pkg)
}
