% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcpkgs-package.R
\docType{package}
\name{srcpkgs-package}
\alias{srcpkgs}
\alias{srcpkgs-package}
\title{srcpkgs: R source packages manager}
\description{
Manage a collection/library of R source packages. Discover, roxygenize, load, test source packages. Enable to use those packages as if they were actually installed. Quickly reload only what is needed on source code change. Run tests and checks in parallel.
}
\section{Features}{


\code{srcpkgs} main objective is to ease development on any project
that uses a collection of R source packages (a library).
It is able to figure out which dependencies are source packages, and is able
to quickly detect changes in any of the used source packages.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kforner/srcpkgs}
  \item Report bugs at \url{https://github.com/kforner/srcpkgs/issues}
}

}
\author{
\strong{Maintainer}: Karl Forner \email{karl.forner@gmail.com} [copyright holder]

}
