% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcpkgs.R
\name{srcpkgs}
\alias{srcpkgs}
\title{creates a new "srcpkgs" object}
\usage{
srcpkgs(pkgs = lapply(paths, devtools::as.package), paths = NULL)
}
\arguments{
\item{pkgs}{an existing srcpkgs object (no op), or a list of
source package-like objects}

\item{paths}{a list of source package paths as a character vector or list}
}
\value{
a \code{srcpkgs} object (a list named after the package names)
}
\description{
creates a new "srcpkgs" object
}
\examples{
# build dummy source packages
pkg1 <- setup_and_get_dummy_srcpkg()
pkg2 <- pkg1
pkg2$package <- "dummy.srcpkg2"

print(srcpkgs(list(pkg1, pkg2)))
print(srcpkgs(paths = pkg1$path))
}
