% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_test.R
\name{summary.pkg_test}
\alias{summary.pkg_test}
\title{summarizes by test source file. Use by=NA to summarize all rows to get an overview}
\usage{
\method{summary}{pkg_test}(object, col = "file", ...)
}
\description{
summarizes by test source file. Use by=NA to summarize all rows to get an overview
}
