% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_r_loaders.R
\name{unhack_r_loaders}
\alias{unhack_r_loaders}
\title{untraces library() and loadNamespace()}
\usage{
unhack_r_loaders()
}
\value{
no return value, called for side-effects
}
\description{
The function is reentrant.
}
\seealso{
\code{\link[=hack_r_loaders]{hack_r_loaders()}}
}
