% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_srcpkgs.R
\name{find_srcpkgs}
\alias{find_srcpkgs}
\title{find available source packages}
\usage{
find_srcpkgs(root = get_project_root(), prune = TRUE)
}
\arguments{
\item{root}{from where to search for source packages}

\item{prune}{whether to report packages contained inside another package (e.g. in tests/)}
}
\value{
a "srcpkgs" object
}
\description{
find available source packages
}
