% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_srcpkgs.R
\name{get_srcpkgs}
\alias{get_srcpkgs}
\title{get the current source packages list, if any, or discover them using \code{\link[=find_srcpkgs]{find_srcpkgs()}}}
\usage{
get_srcpkgs()
}
\value{
the source packages as a "scrpkgs" object
}
\description{
get the current source packages list, if any, or discover them using \code{\link[=find_srcpkgs]{find_srcpkgs()}}
}
