% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{pkg_list_attached}
\alias{pkg_list_attached}
\title{lists the packages that are attached, i.e. present in the R search() path}
\usage{
pkg_list_attached()
}
\value{
the names of attached package name as a character vector
}
\description{
lists the packages that are attached, i.e. present in the R search() path
}
\examples{
print(sort(pkg_list_attached()))
}
