% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_load.R
\name{pkg_load}
\alias{pkg_load}
\title{loads or reloads if needed a source package, taking care of its dependencies}
\usage{
pkg_load(
  pkgid,
  src_pkgs = get_srcpkgs(),
  deps_graph = compute_pkgs_dependencies_graph(src_pkgs),
  attach = TRUE,
  ignore_suggests = TRUE,
  force = FALSE,
  roxygen = TRUE,
  helpers = FALSE,
  export_all = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{pkgid}{a package name, path or object}

\item{src_pkgs}{a collection of source packages as a \code{srckgs} object.}

\item{deps_graph}{a source package dependencies graph, as an igraph object}

\item{attach}{Whether to attach a package environment to the search
path. If \code{FALSE} \code{load_all()} behaves like \code{loadNamespace()}. If
\code{TRUE} (the default), it behaves like \code{library()}. If \code{FALSE},
the \code{export_all}, \code{export_imports}, and \code{helpers} arguments have
no effect.}

\item{ignore_suggests}{whether not to load suggested packages. if FALSE, the suggested are processed
like imports}

\item{force}{if FALSE, will only reload packages if needed, i.e. if some changes are detected}

\item{roxygen}{whether to automatically roxygenise packages (if needed)}

\item{helpers}{if \code{TRUE} loads \pkg{testthat} test helpers.}

\item{export_all}{If \code{TRUE} (the default), export all objects.
If \code{FALSE}, export only the objects that are listed as exports
in the NAMESPACE file.}

\item{quiet}{whether to be quiet/silent}

\item{...}{
  Arguments passed on to \code{\link[devtools:load_all]{devtools::load_all}}
  \describe{
    \item{\code{path}}{Path to a package, or within a package.}
    \item{\code{reset}}{clear package environment and reset file cache before loading
any pieces of the package. This largely equivalent to running
\code{\link[pkgload:unload]{unload()}}, however the old namespaces are not completely removed and no
\code{.onUnload()} hooks are called. Use \code{reset = FALSE} may be faster for
large code bases, but is a significantly less accurate approximation.}
    \item{\code{recompile}}{DEPRECATED. force a recompile of DLL from source code, if
present. This is equivalent to running \code{\link[pkgbuild:clean_dll]{pkgbuild::clean_dll()}} before
\code{load_all}}
  }}
}
\value{
FALSE iff not a single package was (re)loaded
}
\description{
This function will check that all dependent packages are up-to-date, and
document and reload them as needed.
}
\details{
To be able to properly load a package, its dependent source packages
must be loaded in proper order.
i.e. if A-->B-->C, the load order must be C, B, A
}
