% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_load.R
\name{pkg_load}
\alias{pkg_load}
\title{loads or reloads if needed a source package}
\usage{
pkg_load()
}
\value{
TRUE iff pkg was reloaded
}
\description{
This function will check that all dependent packages are up-to-date, and
document and reload them as needed.
}
\details{
To be able to properly load a package, its dependent source packages
must be loaded in proper order.
i.e. if A-->B-->C, the load order must be C, B, A
}
