% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{reset the srcpkgs settings}
\usage{
reset(root = find_project_root(), srcpkgs_paths = find_srcpkgs_paths(root))
}
\arguments{
\item{root}{directory from where to search for source packages}

\item{srcpkgs_paths}{paths to the source packages folders}
}
\description{
reset the srcpkgs settings
}
