% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{resets the \code{srcpkgs} settings}
\usage{
reset(root = find_project_root(), srcpkgs_paths = find_srcpkgs_paths(root))
}
\arguments{
\item{root}{directory from where to search for source packages}

\item{srcpkgs_paths}{paths to the source packages folders}
}
\value{
the settings (cf \code{\link[=settings]{settings()}}) invisibly
}
\description{
With this function, you can reset or set precisely the settings.
}
\examples{
\dontrun{
# reset to appropriate defaults based on your current directory
reset()

# explictly set the project root
reset(root = 'my/project/dir')

# explictly set the source package paths (very unlikely)
reset(srcpkgs_paths = c('pkgs/mypkg1', 'pkgs/mypkg2'))
}
}
